fix: refactor example app architecure
Changed files structure and TryItOutViewModel
Changes:
	renamed:    example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/body/BodyFragment.kt -> example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/BodyFragment.kt
	renamed:    example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/consent/ConsentFragment.kt -> example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/ConsentFragment.kt
	renamed:    example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/destination/DestinationFragment.kt -> example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/DestinationFragment.kt
	renamed:    example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/ownership/OwnershipFragment.kt -> example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/OwnershipFragment.kt
	renamed:    example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/try_it_out/TryItOutReq.kt -> example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/Request.kt
	renamed:    example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/try_it_out/TryItOutAdapter.kt -> example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/TryItOutAdapter.kt
	renamed:    example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/try_it_out/TryItOutFragment.kt -> example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/TryItOutFragment.kt
	renamed:    example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/try_it_out/TryItOutViewHolder.kt -> example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/TryItOutViewHolder.kt
	renamed:    example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/try_it_out/TryItOutViewModel.kt -> example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/TryItOutViewModel.kt
	renamed:    example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/wallet/WalletListAdapter.kt -> example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/WalletListAdapter.kt
	renamed:    example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/wallet/WalletListFragment.kt -> example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/WalletListFragment.kt
	renamed:    example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/wallet/WalletListViewHolder.kt -> example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/WalletListViewHolder.kt
	deleted:    example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/destination/Destination.kt
