+++ example_app/app/src/main/java/com/mytiki/tiki_sdk_android/example_app/WalletListViewHolder.kt
